+++ packages/fe/content/pages/index.json
@@ -1,10 +1,10 @@
 {
   "seo": {
-    "title": "Home 3 — Filecoin Plus",
-    "description": ""
+    "title": "Filecoin Plus",
+    "description": "Apply for DataCap with Filecoin Plus, a social layer mechanism that incentivizes all participants to make Filecoin more useful."
   },
   "og": {
-    "site_name": "Home 3 — Filecoin Plus",
+    "site_name": "Filecoin Plus",
     "url": "",
     "type": "website",
     "image": "/open-graph-image.png"
